# core-lending-asset-finance-sdk/src/main/resources/openapi.yml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:18080
  description: Generated server url
...
paths:
...
  /api/v1/asset-finance-agreements/{agreementId}/assets/{assetId}/service-events/{eventId}:
    get:
      tags:
      - ServiceEvent
      summary: Get a service event by ID
      description: Retrieve a specific service event by its unique identifier
      operationId: getById_4
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: assetId
        in: path
        description: Unique identifier of the asset
        required: true
        schema:
          type: string
          format: uuid
      - name: eventId
        in: path
        description: Unique identifier of the service event
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Service event found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceEventDTO"
        "404":
          description: "Service event, asset, or agreement not found"
        "500":
          description: Internal server error
    put:
      tags:
      - ServiceEvent
      summary: Update an existing service event
      description: Update all fields of an existing service event
      operationId: update_4
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: assetId
        in: path
        description: Unique identifier of the asset
        required: true
        schema:
          type: string
          format: uuid
      - name: eventId
        in: path
        description: Unique identifier of the service event to update
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceEventDTO"
        required: true
      responses:
        "200":
          description: Service event updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceEventDTO"
        "400":
          description: Invalid input data
        "404":
          description: "Service event, asset, or agreement not found"
        "500":
          description: Internal server error
    delete:
      tags:
      - ServiceEvent
      summary: Delete a service event
      description: Permanently delete a service event
      operationId: delete_4
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: assetId
        in: path
        description: Unique identifier of the asset
        required: true
        schema:
          type: string
          format: uuid
      - name: eventId
        in: path
        description: Unique identifier of the service event to delete
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "204":
          description: Service event deleted successfully
        "404":
          description: "Service event, asset, or agreement not found"
        "500":
          description: Internal server error
...
  /api/v1/asset-finance-agreements/{agreementId}/assets/{assetId}/service-events:
    get:
      tags:
      - ServiceEvent
      summary: List/Search service events for an asset
      description: Retrieve a paginated list of service/maintenance events for a specific
        asset
      operationId: findAll_4
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: assetId
        in: path
        description: Unique identifier of the asset
        required: true
        schema:
          type: string
          format: uuid
      - name: filterRequest
        in: query
        description: Filter criteria for searching service events
        required: true
        schema:
          $ref: "#/components/schemas/FilterRequestServiceEventDTO"
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      responses:
        "200":
          description: Successfully retrieved list of service events
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PaginationResponseServiceEventDTO"
        "400":
          description: Invalid filter request
        "404":
          description: Asset or agreement not found
        "500":
          description: Internal server error
    post:
      tags:
      - ServiceEvent
      summary: Create a new service event for an asset
      description: "Create a new service/maintenance event record (maintenance, damage,\
        \ repair, etc.)"
      operationId: create_4
      parameters:
      - name: agreementId
        in: path
        description: Unique identifier of the agreement
        required: true
        schema:
          type: string
          format: uuid
      - name: assetId
        in: path
        description: Unique identifier of the asset
        required: true
        schema:
          type: string
          format: uuid
      - name: X-Idempotency-Key
        in: header
        description: "Unique key for idempotent requests. If provided, ensures that\
          \ identical requests with the same key will only be processed once."
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceEventDTO"
        required: true
      responses:
        "200":
          description: Service event created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceEventDTO"
        "400":
          description: Invalid input data
        "404":
          description: Asset or agreement not found
        "500":
          description: Internal server error
...
components:
  schemas:
...
    ServiceEventDTO:
      required:
      - assetFinanceAssetId
      - eventDate
      - eventType
      type: object
      properties:
        serviceEventId:
          type: string
          format: uuid
          readOnly: true
        assetFinanceAssetId:
          type: string
          format: uuid
        eventDate:
          type: string
          format: date
        eventType:
          type: string
          enum:
          - MAINTENANCE
          - DAMAGE
          - RETURN
          - INSPECTION
          - MODIFICATION
        cost:
          type: number
        note:
          maxLength: 1000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: "The direction of sorting, either ASC or DESC."
          example: DESC
          default: DESC
      description: "Represents a pagination request for retrieving paginated results,\
        \ including page number, size, sort field, and direction."
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: "#/components/schemas/RangeObject"
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
...
    FilterRequestServiceEventDTO:
      required:
      - pagination
      type: object
      properties:
        filters:
          $ref: "#/components/schemas/ServiceEventDTO"
        rangeFilters:
          $ref: "#/components/schemas/RangeFilter"
        pagination:
          $ref: "#/components/schemas/PaginationRequest"
        options:
          $ref: "#/components/schemas/FilterOptions"
      description: Generic filter request that includes both filter criteria and pagination
    PaginationResponseServiceEventDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: "#/components/schemas/ServiceEventDTO"
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: "The current page number, typically zero-based."
          format: int32
      description: Represents a paginated response containing a list of items and
        pagination metadata.
...